rename /app to /amarillo
changed amarillo.app to amarillo in Dockerfile and documentation

changed 'app' to 'amarillo' in readme and tests

added __init__.py to /amarillo
