+++ .gitignore
@@ -141,4 +141,5 @@ data/enhanced/
 data/failed/
 data/trash/
 data/gtfs/
+data/tmp
