+++ README.md
@@ -13,7 +13,7 @@ Create a virtual environment `python3 -m venv venv`.
 
 Activate the environment and install the dependencies `pip install -r requirements.txt`.
 
-Run `uvicorn app.main:app`. 
+Run `uvicorn amarillo.main:app`. 
 
 In development, you can use `--reload`. 
 
@@ -44,23 +44,23 @@ Permissions work this way
 
 ### GTFS-RT python bindings
 
-In case you modify or update the proto-files in app/proto, you'll need to regenerate the python bindings. First, create the python files:
+In case you modify or update the proto-files in amarillo/proto, you'll need to regenerate the python bindings. First, create the python files:
 
 ```sh
-$ cd app/proto
+$ cd amarillo/proto
 $ protoc --version
 libprotoc 3.21.6
 $ protoc --proto_path=. --python_out=../services/gtfsrt gtfs-realtime.proto realtime_extension.proto
-$ sed 's/import gtfs_realtime_pb2/import app.services.gtfsrt.gtfs_realtime_pb2/g' ../services/gtfsrt/realtime_extension_pb2.py | sponge ../services/gtfsrt/realtime_extension_pb2.py
+$ sed 's/import gtfs_realtime_pb2/import amarillo.services.gtfsrt.gtfs_realtime_pb2/g' ../services/gtfsrt/realtime_extension_pb2.py | sponge ../services/gtfsrt/realtime_extension_pb2.py
 ```
 
 ## Testing
 
-In the top directory, run `pytest app/tests`.
+In the top directory, run `pytest amarillo/tests`.
 
 ## Docker
 
 Based on [tiangolo/uvicorn-gunicorn:python3.9-slim](https://github.com/tiangolo/uvicorn-gunicorn-docker)
 
 - build `docker build -t amarillo .`
-- run `docker run --rm --name amarillo -p 8000:80 -e ADMIN_TOKEN=$ADMIN_TOKEN -e RIDE2GO_TOKEN=$RIDE2GO_TOKEN -e TZ=Europe/Berlin -v $(pwd)/data:/app/data amarillo`
+- run `docker run --rm --name amarillo -p 8000:80 -e MODULE_NAME=amarillo.main -e MAX_WORKERS="1" -e ADMIN_TOKEN=$ADMIN_TOKEN -e RIDE2GO_TOKEN=$RIDE2GO_TOKEN -e TZ=Europe/Berlin -v $(pwd)/data:/app/data amarillo`
