+++ amarillo/services/gtfs.py
@@ -1,6 +1,6 @@
-import app.services.gtfsrt.gtfs_realtime_pb2 as gtfs_realtime_pb2
-import app.services.gtfsrt.realtime_extension_pb2 as mfdzrte
-from app.services.gtfs_constants import *
+import amarillo.services.gtfsrt.gtfs_realtime_pb2 as gtfs_realtime_pb2
+import amarillo.services.gtfsrt.realtime_extension_pb2 as mfdzrte
+from amarillo.services.gtfs_constants import *
 from google.protobuf.json_format import MessageToDict
 from google.protobuf.json_format import ParseDict
 from datetime import datetime, timedelta
