+++ amarillo/services/trips.py
@@ -1,10 +1,10 @@
-from app.models.gtfs import GtfsTimeDelta, GtfsStopTime
-from app.models.Carpool import MAX_STOPS_PER_TRIP, Carpool, Weekday, StopTime, PickupDropoffType
-from app.services.config import config
-from app.services.gtfs_constants import *
-from app.services.routing import RoutingService, RoutingException
-from app.services.stops import is_carpooling_stop
-from app.utils.utils import assert_folder_exists, is_older_than_days, yesterday, geodesic_distance_in_m
+from amarillo.services.config import config
+from amarillo.models.gtfs import GtfsTimeDelta, GtfsStopTime
+from amarillo.models.Carpool import MAX_STOPS_PER_TRIP, Carpool, Weekday, StopTime, PickupDropoffType
+from amarillo.services.gtfs_constants import *
+from amarillo.services.routing import RoutingService, RoutingException
+from amarillo.services.stops import is_carpooling_stop
+from amarillo.utils.utils import assert_folder_exists, is_older_than_days, yesterday, geodesic_distance_in_m
 from shapely.geometry import Point, LineString, box
 from geojson_pydantic.geometries import LineString as GeoJSONLineString
 from datetime import datetime, timedelta
