+++ enhancer.py
@@ -5,10 +5,10 @@ import logging.config
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
 
-from app.configuration import configure_enhancer_services
-from app.utils.container import container
-from app.models.Carpool import Carpool
-from app.utils.utils import agency_carpool_ids_from_filename
+from amarillo.configuration import configure_enhancer_services
+from amarillo.utils.container import container
+from amarillo.models.Carpool import Carpool
+from amarillo.utils.utils import agency_carpool_ids_from_filename
 
 logging.config.fileConfig('logging.conf', disable_existing_loggers=False)
 logger = logging.getLogger("enhancer")
